requirements.txt
python-telegram-bot==21.1.1
python-dotenv==1.0.1
requests==2.31.0